Move GetVersionHeight tests from GitExtensionsTests to VersionOracleTests
